Move ObjectsCollection to 'common' directory
